Add method to update Samples
When new document is added to Samples
collection and browser is redirecrted
to its page, allow to update all
parameters and display File Upload
button. When file is uploaded or some
of the fields is changed, call
samples/upsert. Add bwControl input to
BWUploadPanel component to return path
to uploaded file. Add BWInputFile class
for all File type inputs.