Add enroll component
When server is run the function
setExtraUsers is called. The function
creates new users for all emails
listed in extra_users section of
settings.json. Each new user is
created without password, so he cannot
log in. To log in he need to follow the
link sent in enrollment email and reset
the password.