Fix bug in git, update getProjectShares
For Git there was not correct default value for
workflowsDir (extra / at the beginning)

For getProjectShares, if not rc_server logic
were added.

Because all the imported users belong to the
satellite's domain, all the labs have owner
and all the projects have main lab, we can
simplify the logic of sharing the projects
between labs.