Refactore Downloads to support "core:" protocol
For every file that should be downloaded we
need to find the specific module to generate
correct link. We select module based on the
protocol. Protocol should end with ":", for
example `core:`.

To process `core:` protocol the module responsible
for generating correct links should also provide
session id to retreave data without password.
Session id, as long as the correct links for all
available files, are refreshed regularly