Fix missing pie-chart when UI requests data with wron id
If function request_file_data is called with file id that
is missing in the raw_data_files collection, the chain of
observables breaks on the first exception (perhaps it's
related to the mergeMap that is called right after the
function that fails with exception). To prevent it, we
need to check if file is present in collection and in case
of missing file, filter out event with error from the chain.
This bug rarely appears because most of the times UI requests
files based on IDs that he received from BioWardrobe-NG.